Split multi-variant table into several row components
